Move logic for starting server to a startServer file
