+++ src/tests/register.test.ts
@@ -1,11 +1,14 @@
 import { request } from 'graphql-request';
 
-import { host } from "./constants";
-import { User } from '../entity/User';
-import { createTypeormConn } from '../utils/createTypeormConn';
+ import { User } from '../entity/User';
+import { startServer } from '../startServer';
+
+let getHost = () => '';
 
 beforeAll(async () => {
-    await createTypeormConn();
+    const app = await startServer();
+    const { port } = app.address();
+    getHost = () => `http://127.0.0.1:${port}`;
 })
 
 const email = 'fortune@fortune.com';
@@ -17,15 +20,8 @@ const mutation = `
     }
 `;
 
-/**
- * @todo
- * Use test db
- * drop test db after test is done running or is about to start
- * test script should also start server before test starts running
- * Avoid createConnection()
- */
 test('Ensures a user is registered', async() => {
-    const response = await request(host, mutation);
+    const response = await request(getHost(), mutation);
     expect(response).toEqual({ register: true });
     const users = await User.find({ where: { email }});
     const user = users[0];
